Remove old swap classes, fix imports
