--- app/src/main/java/io/horizontalsystems/bankwallet/modules/swap/SwapBaseFragment.kt
@@ -1,36 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.swap
-
-import android.os.Bundle
-import android.view.View
-import androidx.navigation.navGraphViewModels
-import io.horizontalsystems.bankwallet.R
-import io.horizontalsystems.bankwallet.core.BaseFragment
-
-abstract class SwapBaseFragment : BaseFragment() {
-
-    private val mainViewModel by navGraphViewModels<SwapMainViewModel>(R.id.swapFragment)
-    protected val dex: SwapMainModule.Dex
-        get() = mainViewModel.dex
-
-    protected abstract fun restoreProviderState(providerState: SwapMainModule.SwapProviderState)
-    protected abstract fun getProviderState(): SwapMainModule.SwapProviderState
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        restoreProviderState(mainViewModel.providerState)
-    }
-
-    override fun onStop() {
-        super.onStop()
-
-        mainViewModel.providerState = getProviderState()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-
-        mainViewModel.providerState = getProviderState()
-    }
-
-}
