+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapx/SwapTokenService.kt
@@ -6,7 +6,7 @@ import io.horizontalsystems.bankwallet.core.fiat.FiatService
 import io.horizontalsystems.bankwallet.entities.CoinValue
 import io.horizontalsystems.bankwallet.entities.CurrencyValue
 import io.horizontalsystems.bankwallet.modules.send.SendModule
-import io.horizontalsystems.bankwallet.modules.swap.coincard.InputParams
+import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.InputParams
 import io.horizontalsystems.marketkit.models.Token
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.asCoroutineDispatcher
