+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapx/SwapXMainViewModel.kt
@@ -22,24 +22,22 @@ import io.horizontalsystems.bankwallet.core.subscribeIO
 import io.horizontalsystems.bankwallet.entities.Address
 import io.horizontalsystems.bankwallet.modules.evmfee.GasDataError
 import io.horizontalsystems.bankwallet.modules.send.evm.SendEvmData
-import io.horizontalsystems.bankwallet.modules.swap.SwapActionState
-import io.horizontalsystems.bankwallet.modules.swap.SwapButtons
-import io.horizontalsystems.bankwallet.modules.swap.SwapMainModule
-import io.horizontalsystems.bankwallet.modules.swap.SwapViewItemHelper
-import io.horizontalsystems.bankwallet.modules.swap.coincard.SwapCoinCardViewModel
-import io.horizontalsystems.bankwallet.modules.swap.oneinch.OneInchKitHelper
-import io.horizontalsystems.bankwallet.modules.swap.oneinch.OneInchSwapParameters
-import io.horizontalsystems.bankwallet.modules.swap.uniswap.UniswapModule
-import io.horizontalsystems.bankwallet.modules.swap.uniswap.UniswapTradeService
 import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.AmountTypeItem
+import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.ExactType
 import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.ISwapProvider
+import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.PriceImpactLevel
 import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.ProviderTradeData
 import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.ProviderViewItem
+import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.SwapActionState
+import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.SwapButtons
 import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.SwapData
+import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.SwapError
 import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.SwapResultState
+import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.UniswapWarnings
 import io.horizontalsystems.bankwallet.modules.swapx.allowance.SwapAllowanceServiceX
 import io.horizontalsystems.bankwallet.modules.swapx.allowance.SwapPendingAllowanceServiceX
 import io.horizontalsystems.bankwallet.modules.swapx.allowance.SwapPendingAllowanceState
+import io.horizontalsystems.bankwallet.modules.swapx.oneinch.OneInchKitHelper
 import io.horizontalsystems.bankwallet.modules.swapx.oneinch.OneInchTradeXService
 import io.horizontalsystems.bankwallet.modules.swapx.providers.UniswapProvider
 import io.horizontalsystems.bankwallet.modules.swapx.uniswap.UniswapTradeXService
@@ -68,6 +66,7 @@ class SwapXMainViewModel(
     private val currencyManager: CurrencyManager,
     private val adapterManager: IAdapterManager,
 ) : ViewModel() {
+    private val maxValidDecimals = 8
 
     private val disposable = CompositeDisposable()
     private val tradeDisposable = CompositeDisposable()
@@ -86,7 +85,7 @@ class SwapXMainViewModel(
             )
         }
 
-    private var amountType: SwapMainModule.AmountType = SwapMainModule.AmountType.ExactFrom
+    private var exactType: ExactType = ExactType.ExactFrom
     private var balanceFrom: BigDecimal? = null
     private var availableBalance: String? = null
     private var amountTypeSelect = buildAmountTypeSelect()
@@ -272,7 +271,7 @@ class SwapXMainViewModel(
 
                     is SwapData.UniswapData -> {
                         tradeView = uniswapTradeViewItem(swapData, fromTokenService.token, toTokenService.token)
-                        if (amountType == SwapMainModule.AmountType.ExactFrom) {
+                        if (exactType == ExactType.ExactFrom) {
                             toTokenService.onChangeAmount(swapData.data.amountOut.toString(), true)
                         } else {
                             fromTokenService.onChangeAmount(swapData.data.amountIn.toString(), true)
@@ -289,9 +288,9 @@ class SwapXMainViewModel(
                 amountFrom?.let { amountFrom ->
                     if (amountFrom > state.allowance.value) {
                         if (revokeRequired()) {
-                            errors.add(SwapMainModule.SwapError.RevokeAllowanceRequired)
+                            errors.add(SwapError.RevokeAllowanceRequired)
                         } else {
-                            errors.add(SwapMainModule.SwapError.InsufficientAllowance)
+                            errors.add(SwapError.InsufficientAllowance)
                         }
                     }
                 }
@@ -307,7 +306,7 @@ class SwapXMainViewModel(
         amountFrom?.let { amountFrom ->
             val balance = balanceFrom
             if (balance == null || balance < amountFrom) {
-                errors.add(SwapMainModule.SwapError.InsufficientBalanceFrom)
+                errors.add(SwapError.InsufficientBalanceFrom)
             }
         }
 
@@ -318,7 +317,7 @@ class SwapXMainViewModel(
         allErrors = errors
         errorShareService.updateErrors(errors)
 
-        val filtered = allErrors.filter { it !is GasDataError && it !is SwapMainModule.SwapError }
+        val filtered = allErrors.filter { it !is GasDataError && it !is SwapError }
         error = filtered.firstOrNull()?.let { convert(it) }
 
         syncUiState()
@@ -332,7 +331,7 @@ class SwapXMainViewModel(
     }
 
     private fun resyncSwapData() {
-        tradeService.fetchSwapData(fromTokenService.token, toTokenService.token, amountFrom, amountTo, amountType)
+        tradeService.fetchSwapData(fromTokenService.token, toTokenService.token, amountFrom, amountTo, exactType)
     }
 
     private fun syncButtonsState() {
@@ -354,7 +353,7 @@ class SwapXMainViewModel(
 
         tradeService.state is SwapResultState.Ready -> {
             when {
-                allErrors.any { it == SwapMainModule.SwapError.InsufficientBalanceFrom } -> {
+                allErrors.any { it == SwapError.InsufficientBalanceFrom } -> {
                     SwapActionState.Disabled(Translator.getString(R.string.Swap_ErrorInsufficientBalance))
                 }
 
@@ -382,7 +381,7 @@ class SwapXMainViewModel(
             SwapActionState.Disabled(Translator.getString(R.string.Swap_Revoking))
         }
 
-        allErrors.isNotEmpty() && allErrors.all { it == SwapMainModule.SwapError.RevokeAllowanceRequired } -> {
+        allErrors.isNotEmpty() && allErrors.all { it == SwapError.RevokeAllowanceRequired } -> {
             SwapActionState.Enabled(Translator.getString(R.string.Swap_Revoke))
         }
 
@@ -400,11 +399,11 @@ class SwapXMainViewModel(
             SwapActionState.Disabled(Translator.getString(R.string.Swap_Approving), loading = true)
         }
 
-        tradeService.state is SwapResultState.NotReady || allErrors.any { it == SwapMainModule.SwapError.InsufficientBalanceFrom } -> {
+        tradeService.state is SwapResultState.NotReady || allErrors.any { it == SwapError.InsufficientBalanceFrom } -> {
             SwapActionState.Hidden
         }
 
-        allErrors.any { it == SwapMainModule.SwapError.InsufficientAllowance } -> {
+        allErrors.any { it == SwapError.InsufficientAllowance } -> {
             SwapActionState.Enabled(Translator.getString(R.string.Swap_Approve))
         }
 
@@ -457,7 +456,7 @@ class SwapXMainViewModel(
         }
     }
 
-    private fun oneInchTradeViewItem(params: OneInchSwapParameters, tokenFrom: Token?, tokenTo: Token?) = try {
+    private fun oneInchTradeViewItem(params: SwapXMainModule.OneInchSwapParameters, tokenFrom: Token?, tokenTo: Token?) = try {
         val sellPrice = params.amountTo.divide(params.amountFrom, params.tokenFrom.decimals, RoundingMode.HALF_UP).stripTrailingZeros()
         val buyPrice = params.amountFrom.divide(params.amountTo, params.tokenTo.decimals, RoundingMode.HALF_UP).stripTrailingZeros()
         val (primaryPrice, secondaryPrice) = formatter.prices(sellPrice, buyPrice, tokenFrom, tokenTo)
@@ -477,7 +476,7 @@ class SwapXMainViewModel(
             ProviderTradeData.UniswapTradeViewItem(
                 primaryPrice = primaryPrice,
                 secondaryPrice = secondaryPrice,
-                priceImpact = formatter.priceImpactViewItem(swapData, UniswapTradeService.PriceImpactLevel.Warning),
+                priceImpact = formatter.priceImpactViewItem(swapData, PriceImpactLevel.Warning),
                 guaranteedAmount = formatter.guaranteedAmountViewItem(
                     swapData.data,
                     tokenFrom,
@@ -548,7 +547,7 @@ class SwapXMainViewModel(
     fun onSelectFromCoin(token: Token) {
         fromTokenService.onSelectCoin(token)
         syncBalance(balance(token))
-        if (amountType == SwapMainModule.AmountType.ExactTo) {
+        if (exactType == ExactType.ExactTo) {
             fromTokenService.onChangeAmount(null, true)
         }
         if (token == toTokenService.token) {
@@ -562,7 +561,7 @@ class SwapXMainViewModel(
 
     fun onSelectToCoin(token: Token) {
         toTokenService.onSelectCoin(token)
-        if (amountType == SwapMainModule.AmountType.ExactFrom) {
+        if (exactType == ExactType.ExactFrom) {
             toTokenService.onChangeAmount(null, true)
         }
         if (token == fromTokenService.token) {
@@ -573,7 +572,7 @@ class SwapXMainViewModel(
     }
 
     fun onFromAmountChange(amount: String?) {
-        amountType = SwapMainModule.AmountType.ExactFrom
+        exactType = ExactType.ExactFrom
         Log.e("TAG", "onFromAmountChange: ${switchService.amountType}", )
         val coinAmount = fromTokenService.getCoinAmount(amount)
         if (amountsEqual(amountFrom, coinAmount)) return
@@ -585,7 +584,7 @@ class SwapXMainViewModel(
     }
 
     fun onToAmountChange(amount: String?) {
-        amountType = SwapMainModule.AmountType.ExactTo
+        exactType = ExactType.ExactTo
         val coinAmount = toTokenService.getCoinAmount(amount)
         if (amountsEqual(amountTo, coinAmount)) return
         amountTo = coinAmount
@@ -616,7 +615,7 @@ class SwapXMainViewModel(
     }
 
     fun onSetAmountInBalancePercent(percent: Int) {
-        val coinDecimals = fromTokenService.token?.decimals ?: SwapCoinCardViewModel.maxValidDecimals
+        val coinDecimals = fromTokenService.token?.decimals ?: maxValidDecimals
         val percentRatio = BigDecimal.valueOf(percent.toDouble() / 100)
         val coinAmount = balanceFrom?.multiply(percentRatio)?.setScale(coinDecimals, RoundingMode.FLOOR) ?: return
 
@@ -652,8 +651,8 @@ class SwapXMainViewModel(
             price = primaryPrice,
             priceImpact = formatter.priceImpactViewItem(swapData)
         )
-        val warnings: List<Warning> = if (swapData.priceImpactLevel == UniswapTradeService.PriceImpactLevel.Forbidden)
-            listOf(UniswapModule.UniswapWarnings.PriceImpactWarning)
+        val warnings: List<Warning> = if (swapData.priceImpactLevel == PriceImpactLevel.Forbidden)
+            listOf(UniswapWarnings.PriceImpactWarning)
         else
             listOf()
 
@@ -668,4 +667,5 @@ class SwapXMainViewModel(
         tradeService.updateSwapSettings(recipient, slippage, ttl)
         syncSwapDataState()
     }
+
 }
