+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapx/confirmation/oneinch/OneInchSendEvmTransactionService.kt
@@ -9,8 +9,8 @@ import io.horizontalsystems.bankwallet.modules.send.evm.SendEvmData
 import io.horizontalsystems.bankwallet.modules.send.evm.settings.SendEvmSettingsService
 import io.horizontalsystems.bankwallet.modules.sendevmtransaction.ISendEvmTransactionService
 import io.horizontalsystems.bankwallet.modules.sendevmtransaction.SendEvmTransactionService
-import io.horizontalsystems.bankwallet.modules.swap.SwapViewItemHelper
-import io.horizontalsystems.bankwallet.modules.swap.oneinch.OneInchSwapParameters
+import io.horizontalsystems.bankwallet.modules.swapx.SwapViewItemHelper
+import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.OneInchSwapParameters
 import io.horizontalsystems.ethereumkit.models.Address
 import io.reactivex.BackpressureStrategy
 import io.reactivex.Flowable
