+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapx/confirmation/oneinch/OneInchSwapConfirmationFragment.kt
@@ -7,7 +7,7 @@ import io.horizontalsystems.bankwallet.core.AppLogger
 import io.horizontalsystems.bankwallet.modules.evmfee.EvmFeeCellViewModel
 import io.horizontalsystems.bankwallet.modules.send.evm.settings.SendEvmNonceViewModel
 import io.horizontalsystems.bankwallet.modules.sendevmtransaction.SendEvmTransactionViewModel
-import io.horizontalsystems.bankwallet.modules.swap.oneinch.OneInchSwapParameters
+import io.horizontalsystems.bankwallet.modules.swapx.SwapXMainModule.OneInchSwapParameters
 import io.horizontalsystems.bankwallet.modules.swapx.confirmation.BaseSwapConfirmationFragment
 import io.horizontalsystems.marketkit.models.BlockchainType
